requirements.txt
Pillow
flask
flask-cors
Flask-Caching
flask_login
requests
beautifulsoup4
pywebview
lxml
utils
